Add files via upload
In this assignment you will use AJAX to call a rest server on ceclnx01 that provides current status information about ceclnx01.

The REST API has the following url:

http://ceclnx01.cec.miamioh.edu/~campbest/cse383-f19-campbest-public/ajax1/info.php/api/v1/FUNCTION (Links to an external site.)

where FUNCTION is:

loadavg: This returns a json datastructure with output from the current load average.
http://ceclnx01.cec.miamioh.edu/~campbest/cse383-f19-campbest-public/ajax1/info.php/api/v1/loadavg (Links to an external site.)
who: returns a json datastructrure with the current users
http://ceclnx01.cec.miamioh.edu/~campbest/cse383-f19-campbest-public/ajax1/info.php/api/v1/who (Links to an external site.)
df: returns a json datastructure with the current disk utilization
http://ceclnx01.cec.miamioh.edu/~campbest/cse383-f19-campbest-public/ajax1/info.php/api/v1/df (Links to an external site.)
network: returns the current tx and rx byte count of the network
http://ceclnx01.cec.miamioh.edu/~campbest/cse383-f19-campbest-public/ajax1/info.php/api/v1/network (Links to an external site.)
Create a new directory in your git working copy called ajax1 and do all work in there.

Use the code at http://ceclnx01.cec.miamioh.edu/~campbest/cse383-f19-campbest-public/ajax1/info.html (Links to an external site.) and edit the file info.js that will make appropriate ajax calls to the rest api and update the display.

Add javascript code so that all three boxes have updated information.  There are three blocks that need information:
On the left is a list of who is logged in and from what IP they are logged in via.
In the middle is a box listing data from the load function.
In the right is a box listing current network usage.
Add javascript code to update the time.
When the ajax calls fails, record the event in the log, newest event at the top.
There is a video below showing my solution.
Paste links to your code and to the git repository here in canvas.